Bug 50957: Populate navbar from JSON block.
The chapterlist, metadata, and previous/next links are now
populated from the page's corresponding JSON block. A category
on each section of the book (of the form
[[Category:Book:Book Title]]) indicates which Book: JSON block
belongs to the book. It then pulls in the JSON and populates
the navbar from that.

This is a pretty huge commit, and since I'm not just tweaking
someone else's code, I'm going to wait for some reviews before
merging this in.

Bug: 50957
Change-Id: I9c2d77ce649c678ae34343032ef68b2b673dc92e