Bug 51256: Load primary CSS separately from JavaScript.
The CSS was all loading with the JavaScript, which meant there was
a delay between the page loading and the CSS being applied. It also
meant that if JavaScript was disabled, no CSS was displayed. Now the
main page's CSS is loaded, and then the JavaScript and JavaScript-
related CSS is loaded after.

Bug: 51256
Change-Id: I0987bf4cab6a49f0ea12579ae73b6d8306af3aa8