Declare NS_BOOK namespace unconditionally
EventLogging deferred namespace registration to a hook handler as a way of
allowing a configuration variable to specify whether the namespace should be
activated or not on a particular wiki. (MediaWiki needs to run the extension's
main file for the configuration variables to be initialized, so the code at the
top level of EventLogging.php doesn't know yet if the namespace should be
activated, because the site-specific configuration data hasn't loaded yet.)
I hadn't realized it at the time, but waiting for registerHandlers to be called
before registering the namespace means the namespace is registered too late,
leading to bug 45031.

Fortunately, BookManagerv2 needn't go through the trouble. Any wiki that
enables the extension will also want the extension's namespace; there is no
case in which you'd want one but not the other. So it's OK to declare the
namespace unconditionally at the file level in BookManagerv2.php.

This patch moves the registration of the NS_BOOK namespace and its content
model out of the hook handler, to BookManagerv2.php.

With the namespace logic relocated, the only remaining method in JsonHooks.php
was 'onEditFilterMerged'. I figured it was better to move it to
BookManagerv2.hooks.php and thus have the extension's two hooks consolidated in
a single file, so that's what I did.

Bug: 45031
Change-Id: I76cfa1d5a15d2267e44b794164e8b1bc41d50229